Large cleanup of unused methods from twilltestcase.
In addition to deleting unused methods, this also moves ones only used by test_library_templates into that file.
